Add Fields Serializing

+++ editor/scripting/robot2D-ScriptCore/robot2D-ScriptCore/InternalCalls.cs
@@ -6,19 +6,19 @@ namespace robot2D
     public static class InternalCalls
     {
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal extern static bool Entity_HasComponent(ulong entityID, Type componentType);
+        internal extern static bool Entity_HasComponent(int entityID, Type componentType);
         
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal extern static ulong Entity_FindEntityByName(string name);
+        internal extern static int Entity_FindEntityByName(string name);
         
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal extern static object GetScriptInstance(ulong entityID);
+        internal extern static object GetScriptInstance(int entityID);
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal extern static void TransformComponent_GetTranslation(ulong entityID, out Vector2 translation);
+        internal extern static void TransformComponent_GetTranslation(int entityID, out Vector2 translation);
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        internal extern static void TransformComponent_SetTranslation(ulong entityID, ref Vector2 translation);
+        internal extern static void TransformComponent_SetTranslation(int entityID, ref Vector2 translation);
         
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         internal extern static bool Input_IsKeyDown(KeyCode keycode);
